+++ packages/ui/lib/components/index.ts
@@ -3,6 +3,7 @@ export * from "./card-carousel";
 export * from "./card-info";
 export * from "./dialog";
 export * from "./empty-state";
+export * from "./floating-button";
 export * from "./icon";
 export * from "./list-item-action";
 export * from "./logo";